Split BusExtensions and add FuturePublishAsync
