+++ Source/EasyNetQ/BusExtensions.cs
@@ -1,51 +1,11 @@
 ﻿using System;
 using System.Threading.Tasks;
 using EasyNetQ.SystemMessages;
-using EasyNetQ.Topology;
 
 namespace EasyNetQ
 {
     public static class BusExtensions
     {
-        /// <summary>
-        /// Schedule a message to be published at some time in the future.
-        /// </summary>
-        /// <typeparam name="T">The message type</typeparam>
-        /// <param name="bus">The IBus instance to publish on</param>
-        /// <param name="messageDelay">The delay time for message to publish in future</param>
-        /// <param name="message">The message to response with</param>
-        public static void FuturePublish<T>(this IBus bus, TimeSpan messageDelay, T message) where T : class
-        {
-            Preconditions.CheckNotNull(message, "message");
-
-            var advancedBus = bus.Advanced;
-            var conventions = advancedBus.Container.Resolve<IConventions>();
-            var connectionConfiguration = advancedBus.Container.Resolve<IConnectionConfiguration>();
-            var delay = Round(messageDelay);
-            var delayString = delay.ToString(@"hh\_mm\_ss");
-            var exchangeName = conventions.ExchangeNamingConvention(typeof (T));
-            var futureExchangeName = exchangeName + "_" + delayString;
-            var futureQueueName = conventions.QueueNamingConvention(typeof (T), delayString);
-            var futureExchange = advancedBus.ExchangeDeclare(futureExchangeName, ExchangeType.Topic);
-            var futureQueue = advancedBus.QueueDeclare(futureQueueName, perQueueTtl: (int) delay.TotalMilliseconds, deadLetterExchange: exchangeName);
-            advancedBus.Bind(futureExchange, futureQueue, "#");
-            var easyNetQMessage = new Message<T>(message)
-                {
-                    Properties =
-                        {
-                            DeliveryMode = (byte) (connectionConfiguration.PersistentMessages ? 2 : 1)
-                        }
-                };
-
-            bus.Advanced.Publish(futureExchange, "#", false, false, easyNetQMessage);
-        }
-
-        private static TimeSpan Round(TimeSpan timeSpan)
-        {
-            return new TimeSpan(timeSpan.Days, timeSpan.Hours, timeSpan.Minutes, timeSpan.Seconds, 0);
-        }
-
-
         /// <summary>
         /// Schedule a message to be published at some time in the future.
         /// This required the EasyNetQ.Scheduler service to be running.
@@ -77,7 +37,7 @@ namespace EasyNetQ
             var typeNameSerializer = advancedBus.Container.Resolve<ITypeNameSerializer>();
             var serializer = advancedBus.Container.Resolve<ISerializer>();
 
-            var typeName = typeNameSerializer.Serialize(typeof (T));
+            var typeName = typeNameSerializer.Serialize(typeof(T));
             var messageBody = serializer.MessageToBytes(message);
 
             bus.Publish(new ScheduleMe
@@ -133,7 +93,7 @@ namespace EasyNetQ
             var typeNameSerializer = advancedBus.Container.Resolve<ITypeNameSerializer>();
             var serializer = advancedBus.Container.Resolve<ISerializer>();
 
-            var typeName = typeNameSerializer.Serialize(typeof (T));
+            var typeName = typeNameSerializer.Serialize(typeof(T));
             var messageBody = serializer.MessageToBytes(message);
 
             return bus.PublishAsync(new ScheduleMe
